fixing all tests

+++ sources/app/app/Http/Controllers/ContactsController.php
@@ -27,11 +27,19 @@ class ContactsController extends Controller
 
     public function update(Contact $contact)
     {
+        if(request()->user()->isNot($contact->user)) {
+            return response([],403);
+        }
+
         $contact->update($this->validate_data());
     }
 
     public function destroy(Contact $contact)
     {
+        if(request()->user()->isNot($contact->user)) {
+            return response([],403);
+        }
+
         $contact->delete();
     }
